[IMP] uwsgi: timeout The control of timeouts was not given and even badly negelgted, which could have lead to the situation of rather long running go routine.
This is because http package has no default time out at all and no timeout was set.
Furthermore, timeout for tcp and unix fas fixed at 5 seconds.

This commit enables configuring timeout, and has it set in it's standard config to 5 seconds.
Furthermore it constructs a httpClient which has a timeout set and therefore avoids a possibly
long runing gorutine.

Signed-off-by: David Arnold <dar@devco.co>