Make telegraf depend on IP/hostname availability
If systemd-networkd is used then telegraf is enabled
at boot up time and it relies on network-online.target for DHCP
completion. If systemd-networkd is not being used then we assume
dhclient is being used. On such systems telegraf is disabled upon
boot and is started by dhclient-up-hooks

This fix ensures that on systemd enabled systems, telegraf will
always start with proper hostname and never 'localhost'.

This fix does not handle hostname lost post-boot.