Enable extracting tag values from MQTT topics
The MQTT topic as a whole is stored as a tag, but this is far too coarse
too be easily used when utilising the data further down the line.  This
change allows tag values to be extracted from the MQTT topic letting you
store the information provided in the topic in a meaningful way.  This
is especially (only, really) useful when subscribing to topics with
single level wildcards.  Multilevel wildcard are supported, but can
cause errors at runtime.  Metrics that cause an error will be dropped.