Fixed install/remove of telegraf on non-systemd Debian/Ubuntu systems
When having systemd installed but not actively used (so systemd is not PID 1 eg. not running) we have to assume that we run sysv instead. So i’ve changed the check wether systemd is installed (`which systemctl` ex. on line 63 post-install.sh) to wether systemd is actually running (`pidof systemd`). If it’s not running we can safely fallback to the sysv method for install/remove because `pidof systemd` will exit 1 of systemd is not running. This also fixes problems when running Debian Jessie with package `sysvinit-core` installed and used instead of systemd!

I’ve also optimzed the bash scripts a bit, so for example the return codes are not checked after a command has been run (line 28 post-install.sh) but the actual return code is actively checked (`if <command>; then` instead of `<command>; if [ $? -eq 0 ]; then` - for rationale see: https://github.com/koalaman/shellcheck/wiki/SC2181)

Also changed the remove/purge OR (line 30 post-remove.sh) see: https://github.com/koalaman/shellcheck/wiki/SC2166 and made some minor indentation fixes