Generalize the output of the sensors input plugin
This removes the usage of the -u flag, which makes the output of the
sensors command unreliable.

Additionally it tries to reorganize the fields and make them more
readable.

An additional unit field has been added so that different type of
measurements can be filtered (temeprature, voltages, RPM, etc.)

Discussion: https://community.influxdata.com/t/sensors-input-plugin/847