Add Tags Processor
Metrics created by different input plugins can have tags attached to provide
additional information. Those tags may either be configured in the global_tags
section in the telegraf.conf file or may be provided by the respective plugin.
Telegraf provides a filtering mechanism with the *taginclude* property to only
include certain tags. However, this mechanism will also drop all tags, that are
not explicitly listed. This applies to the plugin specific tags as well as the
global tags.

In order to always provide global tags, this processor adds all configured tags
after the tag filtering was applied. Therefore all those tags will always be
present. This is useful in the case, were only some tags provided by an input
plugin are needed, but it is not exactly known, which other tags may be attached
by the input plugin. Using this processor provides the ability to use
*taginclude* for the wanted tags from the input plugin while still retaining
global tags. It eases the configuration as the global tags do not need to be
repeated in every *taginclude*.

This implementation will add the configured tags to all metrics, no filtering is
supported. When adding the tags, values of tags with the same key will be
overwritten by the processor.

Signed-off-by: Karsten Schnitter <k.schnitter@sap.com>