outputs.graphite: Retry sending metrics immediately after reconnect
If writing to Graphite would fail the plugin would reconnect, but not retry
to send the metrics until the next interval. In a situation when the connection
would break before the next interval the metrics would never reach the Graphite
server. Obviously this is a network issue which but Telegraf should handle these
kind of situations better.

This patch retries to send the metrics one time immediately after reconnecting.