ignore redis info when key name contain "replid"
when grab information from multiple redis nodes, some nodes' xx_replid
are sha1 string like '34055853fce268cfc81aadd5686e1dd79b21b5bf', but
some nodes' xx_replid are just
'0000000000000000000000000000000000000000',
when the all zero string get parsed in golang, it's a int 0, however
sha1 string is parsed as a string, so the type difference cause a type
conflict, which lead to drop of point when write to influxdb.
whatsmore the xx_replid section is useless, so drop this section is
harmless.