Makes the unbound plugin reset statistics on read
Per the unbound documentation, the flag `statistics-cumulative: yes|no`
determines whether or not unbound clears its stat counters on a read
via the `stats` call to unbound-control. Since the original plugin is
hardcoded to use the `stats_noreset` command, it makes the configuration
option in the unbound.conf meaningless and provides no way to change it
in telegraf.

This changes to the plugin to calling the `stats` call to fix this problem.