fix snmp overriding of auto-configured table fields
Whenever the snmp plugin was configured with a table with automatic field discovery, if one of those fields was explicitly overridden in the config and the value of is_tag was changed, the field would be duplicated, once as a tag & once as a field.
This change fixes the behavior so that if a field is explicitly configured, automatic table field discovery doesn't touch it.