Fileinfo parser plugin
Fileinfo is an attempt to approach a specific problem with a generic solution. Ie: It's not generic right now. But, it definitely could be.

The vision for the fileinfo plugin is to allow multiple parsers. For example, you want to tail a file, but want to collect metrics about the file you are tailing, perhaps to see growth rate?

Fileinfo is a superset of a filestat plugin, that should allow the users to capture not only filestat information but information from the filename itself if metadata is included in it.