Improve passthrough in prometheus output and input plugins
Prometheus metrics will get mangled when passed to
another telegraf instance using any of the serializers
due to the loss of type. This is makes the histogram
and summary metrics unusable in prometheus.

This change will store prometheus help and type
as metric tags in prometheus input and restore them
in prometheus output.