Cleanup Ceph Input Plugin
With a new found mastery of Go it was time to revisit this plugin and
make the code more idiomatic.  The main thrust is to unmarshal directly
into data structures rather than using type assertions which makes the
code far more readable.  I've also added additional protections in the
form of extended testing.