Add hostname to TLS config for SNI support
x509_cert plugin was creating TLS connection without specifying a
server name for handshake. If there are multiple certificates on the
same host, the first one is received, resulting insecure host warning.

ServerName in tls.Config turns Server Name Indication (SNI) on for the
handshake, which allows virtual hosts on the same server.