Add input plugin for IPVS
This is a first pass at adding an IPVS input plugin. It is quite
rudimentary and has the following limitations:
1. does not support filtering, reports *all* the stats it sees
2. only reports virtual server stats (not real servers)
3. only works on linux (although IPVS support is available on FreeBSD)

Both of these limitations will be addressed in subsequent commits.

NOTE: currently relies on github.com/amoghe/libipvs, however we
should change this to use github.com/viswananda/netlink since that
is well maintained (needs support for IPVS to be added to it).