Extract end-entity certificate data only
The previous code was iterating over each certificate in the chain and over-writing the "fields" variable with each subsequent certificate. This had the effect of only extracting the data from the last certificate in the chain (typically the issuing authority certificate), when actually the user is likely expecting the first (the end-entity certificate).