Force vishvananda/netlink lib to an older version
The IPVS input plugin imports `docker/libnetwork` which in turn vendors
an older version of `vishvananda/netlink`. When using `dep import` to
pull in `docker/libnetwork` the dep tool is able to see this transitive
dependency but it fails to detect the version and ends up bringing in
a newer `vishvananda/netlink` that is incompatible with `docker/libnetwork`.

This patch pegs the `vishvananda/netlink` lib to the older one that
`docker/netlink` expects. We'll also work with docker to bump up its
dependency version so that we can remove this constraint in the near
future.