Add Graphite output Template Map
Add the template map feature to the Graphite output. It allows a user to
select which metric should be serialized by which template either
by metric name of by the specified tag value instead of using only
a single template for every metric.
Unmatched metrics can either be processed by the default template
of be excluded from the output.