Rename testdata dir with double dots
Unfortunately due to a security check in Go modules (as of the latest
release), the existance of '..' in a file path in a repository causes
the module to be rejected when adding it with `go get`:
https://github.com/golang/go/issues/28001#issuecomment-443309655.

This renames the testdata directory with the double dots and augments
the test code to rename it on the fly back to having two dots to get the
appropriate test behavior. This means the tests cannot run on a
read-only filesystem but hopefully that is not a requirement.