Add test for SNMP over TCP connection setup
Add a new test to verify that agents prefixed with tcp: use the correct
transport protocol. As getConnection does the full connection setup, we
need a server listening on the port to allow the test to pass. This
server is therefore also set up in the test.  Also verify that the other
agents still use UDP.