Elasticsearch: save master for each server (Closes: #2650)
Like described in #2650, the current implementation of isMaster was incorrect.
As calls were done concurrently, the isMaster value was prone to a race condition.
Also when multiple elasticsearch clusters were specified, this was broken.

To fix this, a map was added which contains the nodeID and masterID.
So for each node we know which one is master (if nodeID == masterID).

Test data taken from existing pull request.