Support obtaining all build later than max_build_age
This changes the job query to include a depth parameter so the build
information for the job is obtained in the same result as the job
details. I also limit the size of the body to only the fields we're
interested in using the tree parameter.

Before the change, we would get metrics of 20 nodes and 100 builds in
23.56 seconds on average. After this change, we get metrics of 20 nodes
and 2006 builds in 7.76 seconds on average.

We were able to get a 3x decrease in time while increasing the number
of metrics by over 16x.