Run process in own goroutine and use context for cancelation
This moves all code that manages the daemon process into a loop running
in its own goroutine. A cancelation context and a WaitGroup is used to
signal the gorouting to terminate the process when the plugin is
stopped. This removes a race condition between the previous command loop
and the Stop method.