Set headers correctly on influxdb_listener
The ping endpoint in influx returns the version in a HTTP header (eg X-Influxdb-Version: 1.0).  The presence of this header is checked by Springboot influxdb ping() implementation (at quite possibly other clients).  This PR adds this header to ping responses and additionally sets the Content-Type correctly when used with the verbose parameter.