plugins/filestat: Skip missing files
Previously, the filestat plugin would check for a file's existence,
and then try to compute all sorts of things for the missing file,
which failed and resulted in a log message about possible permission
errors.

Now, the plugin just skips the file if it exists. If an error is
encountered probing an existing file, the plugin now logs that error,
too, instead of hazarding a guess at permission problems.