Set RestartSec to help keep service running after a crash
RestartSec defaults to an aggressive 100ms. When Telegraf fails due to an out of memory issue caused by another process on the system we need Telegraf's service launcher to wait for the OOM killer to handle the issue before restarting Telegraf. Without waiting the Telegraf service will crash 5 times in a row in the span of ~1second and pust the service into a more that will no-longer restart on failure.