An entirely new way to convert values, using the built-in go binary stream reader.  This is accomplished by creating a dynamic binary.ByteOrder.  Some unit tests added, but many more test cases are needed.
The configuration specification allows you to specify order for the entire plugin instance, for a specific request, or for a specific field.  In other words, you can set your own default, then override it later.  README.md still needs to be updated to reflect this, and tests for floats haven't been written yet.