Add include_mounts and exclude_mounts options to nfsclient.
The "include_mounts" and "exclude_mounts" filter which entries from
/proc/self/mountstats are actually processed and parsed.  Filtering at
collection time can be important for sites with very large numbers of
NFS mounts.  By default (if neither option is specified), all mounts
are processed.

Both lists are regexp patterns, and are matched against the local NFS
mount point, and not the device being mounted.  If "include_mounts"
is set, all mounts will be ignored, except those matching the given
patterns.  When a mount point matches a pattern in "exclude_mounts"
it will be ignored.

Exclusion takes precendence over inclusion:  if a mount is listed in both
"include_mounts" and "exclude_mounts", it will be ignored.

Mounts can still be filtered after collection via tagpass/tagdrop.