Fix AppVeyor CI build error for Windows.
On windows 'nul' is the equivalent to /dev/null. However, we are
executing the command deriving the current tag within the shell
environment of GNU make on windows. Now in some environments the
shell seems to be unixoid using '/dev/null' instead of 'nul' while
on others it's the other way around. AppVeyor and MSys2 use the
unixoid shells and in turn the line

   tag := $(shell git describe --exact-match --tags 2>$(devnull))

will redirect to a file called 'nul' on those systems.
Followup processes try to cleanup the created file which fails as
'nul' is a special name and cannot be deleted straight-forward in
the native windows environment.

We now write to a regular file and delete it directly. While this
is a bit odd, it should work on any platform. Keep the fingers
crossed... :-)

Thanks reimda for testing!