Update gopsutils to v.2.20.7
Additionally to the version bump of gopsutils we need to update
golang.org/x/sys as gopsutils strongly depends on this package.
We have to exactly use the sys-version (that exact commit) as
directly afterwards a variable (SYS__SYSCALL) is removed which
breaks gopsutils 2.20.7 (which is the latest release currently).