Fix execd processor test
On some occasions ingest part of the test (runCountMultiplierProgram)
does not manage to read all the metrics from stdin that are sent to it
before the process is stopped (and stdin closed) by the test. We can
then observe errors like "Error reading stderr" or "Error reading stdout"
(from the test process). This happens more often when the number of
metrics sent in the test increases.

Even less often, a deadlock might occur when the test does not consume
all the metrics and it waits on acc.Wait() for sync.Cond broadcast that
never arrives as in [1].

To fix this spin a separate goroutine which will check that accumulator
received all the expected metrics and only then stop the count
multiplier process.

[1]: https://app.circleci.com/pipelines/github/SumoLogic/telegraf/39/workflows/4a7fd631-25f9-4f99-8134-fe9ff267fc31/jobs/210