fix: plugins/parsers/influx: avoid ParseError.Error panic
A line ends at `\n` or `\r\n`, but not `\r` on its own. The ParseError.Error
logic was assuming that any carriage return terminates the line.
This caused a crash when the buffer is big, there's a single `\r` character and the
error offset was after that, because it assumed that the offset is always before
the end of the current line (a reasonable assumption as long as the line-termination
logic agrees between the Error method and the parsing code).

This bug found by running `go-fuzz`.