Add queries for collecting query store runtime statistics and query store wait statistics, for Azure SQL DB and MI
Add ability to store in a hashtable a custom value associated with the query result, and use it when the query is run next time.
For query store queries, we store the time when the query was run last, and in the next run we only return the data between the last run time and the current time, rounded down to a multiple of the reporting interval.