Improve SQL connection lifecycle
Instead of opening a new SQL connection for every query, open it for a server, and use it for all queries within this server. We observed that opening a new connection for each query lead to sporadic connection errors in the Telegraf log file when monitoring 100 Azure SQL DBs with all queries enabled.