fix: remove ambiguity on '\v' from line-protocol parser
The line protocol parser allows vertical tab ('\v') in whitespace, but also
allows it in measurements, tags keys, tag values, and field keys. This
ambiguity causes a blowup in the parsing state machine and triggers undefined
behaviour when the vertical tab character is seen. The parser will attempt to
simultaneously extend a compoent of the line, and move on to the next one.

This patch removes vertical tab from measurements, tags, and fields. The
resulting state machine goes from 745 states to 90, with a similar reduction in
the number of lines of generated code.