procstat: option to select which metrics collect
If we want to collect only the metric procstat_lookup or procstat_tcp is
a big overhead having to parse lots of files in /proc to get number of
fds, CPU time, memory usage, etc.

Added a new configuration option to be able to select which metrics to
collect.
If the configuration parameter, 'metrics_include', is undefined, the
plugin behaves like before adding this parameter.

The difference (measured with ~1100 procs) between gathering the
procstat_tcp metrics alone or with the default collection of metrics is
around one tenth (~50ms alone, ~500ms with default collection).