fix: prevent goroutine leak in Prometheus input
When configured with `monitor_kubernetes_pods` and `cluster` scope, the
prometheus input will fail to scrape any pods and incur a memory leak.

The reason for failing to scrape any pods is straightforward: it fails
to read any Pod information when getting results from the watch
endpoint, since `pod` is emptied prior to testing for the presence of an
annotation:

```
for event := range watcher.ResultChan() {
        pod = &corev1.Pod{}
        // If the pod is not "ready", there will be no ip associated with it.
        if pod.Annotations["prometheus.io/scrape"] != "true" ||
                !podReady(pod.Status.ContainerStatuses) {
                continue
        }
```

The memory leak arises because we call `watchPod()` expecting a blocking
call, whereas it spawns a goroutine and returns immediately. In the
absence of error, the outer loop ends up retrying every second, leaking
goroutines as it goes.