feat(prometheus): filter pod annotations and labels from metric tags
This commit allows a user to configure a filter to control what pod
annotations and labels to use as metric tags. The primary use case is to
restrain cardinality, since Kubernetes promotes many uses for labels and
annotations which are not relevant for the purposes of analyzing time
series data.