fix: procstat missing tags in procstat_lookup metric
In #9488 the way that tags were built for procstat_lookup was changed
and it was only including the pid_finder and result tags. This is not
consistent with the documentation and is a regression from how they were
previously constructed.

Becuase of the large change to how procstat metrics are gathered, this
will use one of the process metric's tags as a basis for the tags for
procstat_lookup.

Resolves: #9793